+++ buildpacks-jvm-shared/src/lib.rs
@@ -1,0 +1,29 @@
+// Enable rustc and Clippy lints that are disabled by default.
+// https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-crate-dependencies
+#![warn(unused_crate_dependencies)]
+// https://rust-lang.github.io/rust-clippy/stable/index.html
+#![warn(clippy::pedantic)]
+// This lint is too noisy and enforces a style that reduces readability in many cases.
+#![allow(clippy::module_name_repetitions)]
+
+use indoc::formatdoc;
+use libherokubuildpack::log::log_error;
+use std::fmt::Debug;
+
+pub fn log_please_try_again_error<H: AsRef<str>, M: AsRef<str>, E: Debug>(
+    header: H,
+    message: M,
+    error: E,
+) {
+    log_error(
+        header,
+        formatdoc! {"
+            {message}
+
+            Please try again. If this error persists, please contact us:
+            https://help.heroku.com/
+
+            Details: {error:?}
+        ", message = message.as_ref(), error = error },
+    );
+}