+++ buildpacks/jvm-function-invoker/src/layers/mod.rs
@@ -1,3 +1,3 @@
-pub mod bundle;
-pub mod opt;
-pub mod runtime;
+pub(crate) mod bundle;
+pub(crate) mod opt;
+pub(crate) mod runtime;