+++ buildpacks/jvm/src/main.rs
@@ -1,4 +1,6 @@
 // Enable rustc and Clippy lints that are disabled by default.
+// https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-crate-dependencies
+#![warn(unused_crate_dependencies)]
 // https://rust-lang.github.io/rust-clippy/stable/index.html
 #![warn(clippy::pedantic)]
 // This lint is too noisy and enforces a style that reduces readability in many cases.
@@ -10,27 +12,34 @@ mod layers;
 mod util;
 mod version;
 
+use crate::constants::SKIP_HEROKU_JVM_METRICS_AGENT_INSTALLATION_ENV_VAR_NAME;
 use crate::errors::on_error_jvm_buildpack;
 use crate::layers::heroku_metrics_agent::HerokuMetricsAgentLayer;
 use crate::layers::openjdk::OpenJdkLayer;
 use crate::layers::runtime::RuntimeLayer;
-use crate::util::ValidateSha256Error;
+use crate::util::{boolean_buildpack_config_env_var, ValidateSha256Error};
 use crate::version::{NormalizeVersionStringError, ReadVersionStringError};
-pub use constants::*;
+pub(crate) use constants::{
+    JAVA_TOOL_OPTIONS_ENV_VAR_DELIMITER, JAVA_TOOL_OPTIONS_ENV_VAR_NAME, JDK_OVERLAY_DIR_NAME,
+};
 use libcnb::build::{BuildContext, BuildResult, BuildResultBuilder};
-use libcnb::buildpack_main;
 use libcnb::data::build_plan::BuildPlanBuilder;
 use libcnb::data::layer_name;
 use libcnb::detect::{DetectContext, DetectResult, DetectResultBuilder};
 use libcnb::generic::GenericPlatform;
 use libcnb::Buildpack;
+use libcnb::{buildpack_main, Platform};
+#[cfg(test)]
+use libcnb_test as _;
 use libherokubuildpack::download::DownloadError;
 use serde::{Deserialize, Serialize};
+// Work around unused_crate_dependencies. url is used in heroku_database_env_var_rewrite.
+use url as _;
 
-pub struct OpenJdkBuildpack;
+pub(crate) struct OpenJdkBuildpack;
 
 #[derive(Debug)]
-pub enum OpenJdkBuildpackError {
+pub(crate) enum OpenJdkBuildpackError {
     OpenJdkDownloadError(DownloadError),
     MetricsAgentDownloadError(DownloadError),
     MetricsAgentSha256ValidationError(ValidateSha256Error),
@@ -82,7 +91,19 @@ impl Buildpack for OpenJdkBuildpack {
             },
         )?;
 
-        context.handle_layer(layer_name!("heroku_metrics_agent"), HerokuMetricsAgentLayer)?;
+        libherokubuildpack::log::log_header("Installing Heroku JVM metrics agent");
+
+        if boolean_buildpack_config_env_var(
+            context.platform.env(),
+            SKIP_HEROKU_JVM_METRICS_AGENT_INSTALLATION_ENV_VAR_NAME,
+        ) {
+            libherokubuildpack::log::log_info(format!(
+                "Skipping agent installation, {SKIP_HEROKU_JVM_METRICS_AGENT_INSTALLATION_ENV_VAR_NAME} environment variable is set to a truthy value."
+            ));
+        } else {
+            context.handle_layer(layer_name!("heroku_metrics_agent"), HerokuMetricsAgentLayer)?;
+        }
+
         context.handle_layer(layer_name!("runtime"), RuntimeLayer)?;
 
         BuildResultBuilder::new().build()
@@ -94,13 +115,13 @@ impl Buildpack for OpenJdkBuildpack {
 }
 
 #[derive(Deserialize, Debug)]
-pub struct OpenJdkBuildpackMetadata {
+pub(crate) struct OpenJdkBuildpackMetadata {
     #[serde(rename = "heroku-metrics-agent")]
     heroku_metrics_agent: HerokuMetricsAgentMetadata,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq)]
-pub struct HerokuMetricsAgentMetadata {
+pub(crate) struct HerokuMetricsAgentMetadata {
     url: String,
     sha256: String,
 }