+++ buildpacks/maven/src/main.rs
@@ -1,4 +1,6 @@
 // Enable rustc and Clippy lints that are disabled by default.
+// https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-crate-dependencies
+#![warn(unused_crate_dependencies)]
 // https://rust-lang.github.io/rust-clippy/stable/index.html
 #![warn(clippy::pedantic)]
 // This lint is too noisy and enforces a style that reduces readability in many cases.
@@ -19,6 +21,8 @@ use libcnb::detect::{DetectContext, DetectResult, DetectResultBuilder};
 use libcnb::generic::GenericPlatform;
 use libcnb::layer_env::Scope;
 use libcnb::{buildpack_main, Buildpack, Env, Error, Platform};
+#[cfg(test)]
+use libcnb_test as _;
 use libherokubuildpack::download::DownloadError;
 use libherokubuildpack::log::{log_header, log_info};
 use serde::{Deserialize, Serialize};
@@ -28,6 +32,8 @@ use std::fs::Permissions;
 use std::os::unix::fs::PermissionsExt;
 use std::path::{Path, PathBuf};
 use std::process::{Command, ExitStatus};
+#[cfg(test)]
+use ureq as _;
 
 mod errors;
 mod framework;
@@ -37,10 +43,10 @@ mod settings;
 mod util;
 mod warnings;
 
-pub struct MavenBuildpack;
+pub(crate) struct MavenBuildpack;
 
 #[derive(Debug)]
-pub enum MavenBuildpackError {
+pub(crate) enum MavenBuildpackError {
     UnsupportedMavenVersion(String),
     MavenTarballDownloadError(DownloadError),
     MavenTarballSha256IoError(std::io::Error),
@@ -61,14 +67,14 @@ pub enum MavenBuildpackError {
 }
 
 #[derive(Debug, Deserialize)]
-pub struct MavenBuildpackMetadata {
+pub(crate) struct MavenBuildpackMetadata {
     #[serde(rename = "default-version")]
     default_version: String,
     tarballs: HashMap<String, Tarball>,
 }
 
 #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
-pub struct Tarball {
+pub(crate) struct Tarball {
     url: String,
     sha256: String,
 }
@@ -278,7 +284,7 @@ impl From<MavenBuildpackError> for libcnb::Error<MavenBuildpackError> {
     }
 }
 
-pub fn app_dependency_list_path<P: AsRef<Path>>(app_dir: P) -> PathBuf {
+pub(crate) fn app_dependency_list_path<P: AsRef<Path>>(app_dir: P) -> PathBuf {
     app_dir.as_ref().join("target/mvn-dependency-list.log")
 }