+++ buildpacks/maven/tests/misc_tests.rs
@@ -29,13 +29,13 @@ fn mvn_dependency_list() {
             indoc! {"
 
                 The following files have been resolved:
-                   io.undertow:undertow-core:jar:2.2.15.Final:compile
-                   org.jboss.logging:jboss-logging:jar:3.4.1.Final:compile
-                   org.jboss.xnio:xnio-api:jar:3.8.6.Final:compile
+                   io.undertow:undertow-core:jar:2.3.5.Final:compile
+                   org.jboss.logging:jboss-logging:jar:3.4.3.Final:compile
+                   org.jboss.xnio:xnio-api:jar:3.8.8.Final:compile
                    org.wildfly.common:wildfly-common:jar:1.5.4.Final:compile
                    org.wildfly.client:wildfly-client-config:jar:1.0.1.Final:compile
-                   org.jboss.xnio:xnio-nio:jar:3.8.6.Final:runtime
-                   org.jboss.threads:jboss-threads:jar:3.1.0.Final:compile
+                   org.jboss.xnio:xnio-nio:jar:3.8.8.Final:runtime
+                   org.jboss.threads:jboss-threads:jar:3.5.0.Final:compile
                    com.google.guava:guava:jar:30.0-jre:compile
                    com.google.guava:failureaccess:jar:1.0.1:compile
                    com.google.guava:listenablefuture:jar:9999.0-empty-to-avoid-conflict-with-guava:compile
@@ -67,6 +67,7 @@ fn no_unexpected_files_in_app_dir() {
                 /workspace/pom.xml
                 /workspace/src/main/java/com/heroku/App.java
                 /workspace/src/test/java/com/heroku/AppTest.java
+                /workspace/system.properties
                 /workspace/target/classes/com/heroku/App$1.class
                 /workspace/target/classes/com/heroku/App.class
                 /workspace/target/dependency/checker-qual-3.5.0.jar
@@ -75,16 +76,16 @@ fn no_unexpected_files_in_app_dir() {
                 /workspace/target/dependency/guava-30.0-jre.jar
                 /workspace/target/dependency/hamcrest-core-1.3.jar
                 /workspace/target/dependency/j2objc-annotations-1.3.jar
-                /workspace/target/dependency/jboss-logging-3.4.1.Final.jar
-                /workspace/target/dependency/jboss-threads-3.1.0.Final.jar
+                /workspace/target/dependency/jboss-logging-3.4.3.Final.jar
+                /workspace/target/dependency/jboss-threads-3.5.0.Final.jar
                 /workspace/target/dependency/jsr305-3.0.2.jar
                 /workspace/target/dependency/junit-4.13.1.jar
                 /workspace/target/dependency/listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.jar
-                /workspace/target/dependency/undertow-core-2.2.15.Final.jar
+                /workspace/target/dependency/undertow-core-2.3.5.Final.jar
                 /workspace/target/dependency/wildfly-client-config-1.0.1.Final.jar
                 /workspace/target/dependency/wildfly-common-1.5.4.Final.jar
-                /workspace/target/dependency/xnio-api-3.8.6.Final.jar
-                /workspace/target/dependency/xnio-nio-3.8.6.Final.jar
+                /workspace/target/dependency/xnio-api-3.8.8.Final.jar
+                /workspace/target/dependency/xnio-nio-3.8.8.Final.jar
                 /workspace/target/maven-archiver/pom.properties
                 /workspace/target/maven-status/maven-compiler-plugin/compile/default-compile/createdFiles.lst
                 /workspace/target/maven-status/maven-compiler-plugin/compile/default-compile/inputFiles.lst