+++ meta-buildpacks/java/CHANGELOG.md
@@ -4,6 +4,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.8] 2023/03/31
+* Upgraded `heroku/jvm` to `1.0.8`
+
+## [0.6.7] 2023/03/23
+* Upgraded `heroku/jvm` to `1.0.7`
+
 ## [0.6.6] 2023/01/18
 * Upgraded `heroku/jvm` to `1.0.6`